refactor(cli): extract the source-independent semantic-model migration core
Pulls the parts that must not diverge between per-source semantic-model
governance migrations into semantic_model_migration_common: the governance
aspect allowlist, the tag/term union semantics, the editableSchemaMetadata
merge that preserves descriptions, per-entity error isolation, soft-delete
and inbound-reference probes, and the report.

Deliberately narrower than a full strategy abstraction. The two sources'
destinations differ in kind -- Snowflake maps a legacy dataset onto a
semanticModel, while dbt must map it onto another dataset, because a dbt
semanticModel is project-scoped and shared by every semantic model in the
project -- so URN algebra and orchestration stay per-source. Forcing both
through one interface would have added indirection without removing much
code, and would have rewritten all 24 test call sites that bind to the
Snowflake-parameterized signatures.

Snowflake's public surface is unchanged: the shared names are re-imported
into its namespace and its wrappers keep their existing signatures. All 122
migration tests pass with no test file modified, which is what makes this
reviewable as a move.

Two Snowflake-specific pieces are noted as staying put: the synthetic
DIMENSION/FACT/METRIC tag set (dbt has none -- it encodes classification in
nativeDataType) and the uppercase field-path folding (dbt's legacy and new
field paths are identical).
